# bondly-api/docs/swagger.yaml
basePath: /
definitions:
...
  models.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      custody_wallet_address:
        type: string
      email:
        type: string
      encrypted_private_key:
        type: string
      id:
        type: integer
      last_login_at:
        type: string
      nickname:
        type: string
      reputation_score:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      wallet_address:
        type: string
    type: object
  models.WalletBinding:
    properties:
      created_at:
        type: string
      id:
        type: integer
      network:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      wallet_address:
        type: string
    type: object
...
  response.ResponseAny:
    properties:
      code:
        example: 1000
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  response.ResponseWalletBinding:
    properties:
      code:
        example: 1000
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@bondly.com
    name: Bondly API Support
    url: https://bondly.com/support
  description: 这是Bondly项目的API文档，提供用户认证、区块链交互、内容管理和治理功能。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://bondly.com/terms
  title: Bondly API
  version: "1.0"
paths:
...
  /api/v1/wallet-bindings:
    get:
      consumes:
      - application/json
      description: 分页获取钱包绑定列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseWalletBinding'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取钱包绑定列表
      tags:
      - 钱包绑定管理
    post:
      consumes:
      - application/json
      description: 创建新的钱包绑定
      parameters:
      - description: 钱包绑定信息
        in: body
        name: binding
        required: true
        schema:
          $ref: '#/definitions/models.WalletBinding'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ResponseWalletBinding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 创建钱包绑定
      tags:
      - 钱包绑定管理
  /api/v1/wallet-bindings/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的钱包绑定
      parameters:
      - description: 钱包绑定ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 删除钱包绑定
      tags:
      - 钱包绑定管理
    get:
      consumes:
      - application/json
      description: 根据ID获取钱包绑定详情
      parameters:
      - description: 钱包绑定ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseWalletBinding'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取钱包绑定详情
      tags:
      - 钱包绑定管理
    put:
      consumes:
      - application/json
      description: 更新指定钱包绑定的信息
      parameters:
      - description: 钱包绑定ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的钱包绑定信息
        in: body
        name: binding
        required: true
        schema:
          $ref: '#/definitions/models.WalletBinding'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseWalletBinding'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 更新钱包绑定
      tags:
      - 钱包绑定管理
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"